requirements.txt
flask==3.0.2
flask-cors==4.0.0
uvicorn==0.23.2
pydantic==1.10.13
python-dotenv==1.0.0
nltk==3.8.1
textblob==0.17.1
